Seshat/requirements.txt
PyYAML==5.4.1
Seshat==0.3.1
